speed & curvature maybe complete

+++ CurvatureControlforLejosNxj/src/Common/CurvatureControlMethod.java
@@ -4,7 +4,7 @@ public abstract class CurvatureControlMethod {
 
 	public CurvatureControlMethodParm curvatureControlMethodParm;
 
-	public abstract float calcCurvatureCtrlVal(float targCurvature,float curvature,int time);
+	public abstract int calcCurvatureCtrlVal(float targCurvature,float curvature,int time);
 	public abstract void initialize();
 
 	public void setCurvatureCtrlParm(CurvatureControlMethodParm parm){
